% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivebayes.R
\name{naivebayes_classifier}
\alias{naivebayes_classifier}
\title{R6 class for Naive Bayes classifier Object}
\description{
Create an object of class NaiveBayesClassifier.
The Naives Bayes Classifier is a classifier based on applying Bayes' theorem with strong(naive) independance
assumption between the features
}
\details{
Additional details...
}
\references{
references
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NaiveBayesClassifier-new}{\code{naivebayes_classifier$new()}}
\item \href{#method-NaiveBayesClassifier-fit}{\code{naivebayes_classifier$fit()}}
\item \href{#method-NaiveBayesClassifier-predict}{\code{naivebayes_classifier$predict()}}
\item \href{#method-NaiveBayesClassifier-predict_proba}{\code{naivebayes_classifier$predict_proba()}}
\item \href{#method-NaiveBayesClassifier-clone}{\code{naivebayes_classifier$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-new"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naivebayes_classifier$new(alpha = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{(`float`)\cr
Smoothing parameters. Default is 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-fit"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naivebayes_classifier$fit(X, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-predict"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naivebayes_classifier$predict(X)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naivebayes_classifier$predict_proba()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naivebayes_classifier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
